Products Items Screen

+++ feature/products/details/api/src/main/java/com/plana/products/feature/products/details/api/ProductDetailsNavKey.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class ProductDetailsNavKey(
-    val productID: Long
-): NavKey
+    val productID: Int
+) : NavKey
