+++ feature/products/details/impl/src/main/java/com/plana/products/feature/products/details/impl/ProductDetailsScreen.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 
 @Composable
-fun ProductDetailsScreen(productID: Long){
+fun ProductDetailsScreen(productID: Int) {
     Box(
         modifier = Modifier.fillMaxSize(),
         contentAlignment = Alignment.Center
